requirements.txt
openpyxl
langchain-classic
langchain-community
langchain-core
pandas
numpy
matplotlib
reportlab

langchain
langchain_gigachat
langgraph
yandexcloud
deepseek

yandex-ai-studio-sdk
bs4

python-dotenv

python-telegram-bot==21.10